LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-in-cross-origin-iframe-001.sub-expected.txt



FAIL A below-viewport loading=lazy image in a cross origin iframe loads only
       when scrolled into viewport assert_equals: The loading=lazy image should not block the iframe's load event expected "window_loaded" but got "image_loaded"
